+++ examples/llm/__init__.py
@@ -4,16 +4,16 @@
 try:
     import torch
 
+    from examples.llm.src.layers.attention import (FlashCausalAttention,
+                                                   TorchCausalAttention,
+                                                   TritonFlashCausalAttention,
+                                                   alibi_bias)
+    from examples.llm.src.layers.flash_attention import FlashAttention, FlashMHA
     from examples.llm.src.model_registry import COMPOSER_MODEL_REGISTRY
-    from examples.llm.src.models.flash_attention import FlashAttention, FlashMHA
     from examples.llm.src.models.hf import (ComposerHFCausalLM,
                                             ComposerHFPrefixLM, ComposerHFT5)
     from examples.llm.src.models.mosaic_gpt import (GPTMLP, ComposerMosaicGPT,
-                                                    FlashCausalAttention,
-                                                    GPTBlock, MosaicGPT,
-                                                    TorchCausalAttention,
-                                                    TritonFlashCausalAttention,
-                                                    alibi_bias)
+                                                    GPTBlock, MosaicGPT)
     from examples.llm.src.tokenizer import (TOKENIZER_REGISTRY, HFTokenizer,
                                             LLMTokenizer)
 except ImportError as e:
